moved monitor, mouseplatform from ephys rig to device
