+++ src/aind_data_schema/device.py
@@ -318,3 +318,62 @@ class Laser(Device):
     item_number: Optional[str] = Field(None, title="Item number")
     calibration_data: Optional[str] = Field(None, description="Path to calibration data", title="Calibration data")
     calibration_date: Optional[datetime] = Field(None, title="Calibration date")
+
+class MousePlatform(Device):
+    """Description of a mouse platform"""
+
+    surface_material: Optional[str] = Field(None, title="Surface material")
+    date_surface_replaced: Optional[datetime] = Field(None, title="Date surface replaced")
+
+
+class Disc(MousePlatform):
+    """Description of a running disc"""
+
+    platform_type: str = Field("Disc", title="Platform type", const=True)
+    radius: float = Field(..., title="Radius (cm)", units="cm", ge=0)
+    radius_unit: SizeUnit = Field(SizeUnit.CM, title="radius unit")
+
+
+class Tube(MousePlatform):
+    """Description of a tube platform"""
+
+    platform_type: str = Field("Tube", title="Platform type", const=True)
+    diameter: float = Field(..., title="Diameter", ge=0)
+    diameter_unit: SizeUnit = Field(SizeUnit.CM, title="Diameter unit")
+
+
+class Treadmill(MousePlatform):
+    """Description of treadmill platform"""
+
+    platform_type: str = Field("Treadmill", title="Platform type", const=True)
+    treadmill_width: float = Field(..., title="Width of treadmill (mm)", units="mm")
+    width_unit: SizeUnit = Field(SizeUnit.CM, title="Width unit")
+
+class Monitor(Device):
+    """Visual display"""
+
+    # required fields
+    manufacturer: Literal[Manufacturer.LG.value]
+    refresh_rate: int = Field(..., title="Refresh rate (Hz)", units="Hz", ge=60)
+    width: int = Field(..., title="Width (pixels)", units="pixels")
+    height: int = Field(..., title="Height (pixels)", units="pixels")
+    size_unit: SizeUnit = Field(SizeUnit.PX, title="Size unit")
+    viewing_distance: float = Field(..., title="Viewing distance (cm)", units="cm")
+    viewing_distance_unit: SizeUnit = Field(SizeUnit.CM, title="Viewing distance unit")
+
+    # optional fields
+    contrast: Optional[int] = Field(
+        ...,
+        description="Monitor's contrast setting",
+        title="Contrast",
+        ge=0,
+        le=100,
+    )
+    brightness: Optional[int] = Field(
+        ...,
+        description="Monitor's brightness setting",
+        title="Brightness",
+        ge=0,
+        le=100,
+    )
+    position: Optional[RelativePosition] = Field(None, title="Relative position of the monitor")
